Add tests for large PDF handling and improve file reading
- Introduced a new test in `LocalTests.cs` for creating and reading PDFs larger than two gigabytes using `LargePdfTestDocumentGenerator`.
- Updated `FileHeaderOffset.cs` to change the `Value` type from `int` to `long` for better handling of large values.
- Modified `PdfDocumentFactory.cs` to use `FileStream` for opening PDFs, enhancing memory efficiency.
- Added `PdfDocumentOpenTests` class to verify opening of large PDFs with a sparse XRef offset.

+++ src/UglyToad.PdfPig/Parser/FileStructure/FileHeaderOffset.cs
@@ -4,7 +4,7 @@
 /// How many bytes precede the "%PDF-" version header in the file. In some files this 'junk' can
 /// offset all following offset bytes.
 /// </summary>
-internal readonly record struct FileHeaderOffset(int Value) : IEquatable<FileHeaderOffset>
+internal readonly record struct FileHeaderOffset(long Value) : IEquatable<FileHeaderOffset>
 {
     public override string ToString() => Value.ToString();
